# api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/locations/{locationId}/tickets": {
      "post": {
        "operationId": "TicketController_create",
        "parameters": [
          {
            "name": "locationId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTicketDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TicketDto"
                }
              }
            }
          },
          "400": {
            "description": "VALIDATION_ERROR: the request is invalid",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "UNAUTHENTICATED: no valid session",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "UNAUTHORIZED: signed in, but not allowed to do this",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "NOT_FOUND: the resource doesn't exist, or isn't visible to you",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "429": {
            "description": "RATE_LIMITED: too many requests, retry after the Retry-After header's seconds",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          }
        },
        "summary": "Create a ticket (with an inline customer or an existing customerId)",
        "tags": [
          "Tickets"
        ]
      },
      "get": {
        "operationId": "TicketController_list",
        "parameters": [
          {
            "name": "locationId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "q",
            "required": false,
            "in": "query",
            "schema": {
              "maxLength": 100,
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "state",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "open",
                "ended"
              ]
            }
          },
          {
            "name": "createdFrom",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "createdTo",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "createdBy",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "customerId",
            "required": false,
            "in": "query",
            "schema": {
              "pattern": "^\\d+$",
              "type": "string"
            }
          },
          {
            "name": "take",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "type": "number"
            }
          },
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TicketPageDto"
                }
              }
            }
          },
          "400": {
            "description": "VALIDATION_ERROR: the request is invalid",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "UNAUTHENTICATED: no valid session",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "UNAUTHORIZED: signed in, but not allowed to do this",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "NOT_FOUND: the resource doesn't exist, or isn't visible to you",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "429": {
            "description": "RATE_LIMITED: too many requests, retry after the Retry-After header's seconds",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          }
        },
        "summary": "List tickets for a location, newest first",
        "tags": [
          "Tickets"
        ]
      }
    },
    "/locations/{locationId}/tickets/{ticketId}": {
      "get": {
        "operationId": "TicketController_findOne",
        "parameters": [
          {
            "name": "locationId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "ticketId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TicketDetailDto"
                }
              }
            }
          },
          "401": {
            "description": "UNAUTHENTICATED: no valid session",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "UNAUTHORIZED: signed in, but not allowed to do this",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "NOT_FOUND: the resource doesn't exist, or isn't visible to you",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "429": {
            "description": "RATE_LIMITED: too many requests, retry after the Retry-After header's seconds",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          }
        },
        "summary": "Get a ticket's detail with its status timeline",
        "tags": [
          "Tickets"
        ]
      },
      "patch": {
        "operationId": "TicketController_update",
        "parameters": [
          {
            "name": "locationId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "ticketId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTicketDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TicketDto"
                }
              }
            }
          },
          "400": {
            "description": "VALIDATION_ERROR: the request is invalid",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "UNAUTHENTICATED: no valid session",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "UNAUTHORIZED: signed in, but not allowed to do this",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "NOT_FOUND: the resource doesn't exist, or isn't visible to you",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "429": {
            "description": "RATE_LIMITED: too many requests, retry after the Retry-After header's seconds",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          }
        },
        "summary": "Edit a ticket's title/description",
        "tags": [
          "Tickets"
        ]
      }
    },
    "/locations/{locationId}/tickets/{ticketId}/status": {
      "patch": {
        "operationId": "TicketController_updateStatus",
        "parameters": [
          {
            "name": "locationId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "ticketId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTicketStatusDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TicketDto"
                }
              }
            }
          },
          "400": {
            "description": "VALIDATION_ERROR: the request is invalid",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "UNAUTHENTICATED: no valid session",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "UNAUTHORIZED: signed in, but not allowed to do this",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "NOT_FOUND: the resource doesn't exist, or isn't visible to you",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "409": {
            "description": "CONFLICT: not possible in the resource's current state",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "429": {
            "description": "RATE_LIMITED: too many requests, retry after the Retry-After header's seconds",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          }
        },
        "summary": "Move a ticket to another step of its workflow, following ADR 0016's rules",
        "tags": [
          "Tickets"
        ]
      }
    },
    "/locations/{locationId}/tickets/{ticketId}/resend-link": {
      "post": {
        "operationId": "TicketController_resendTrackingLink",
        "parameters": [
          {
            "name": "locationId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "ticketId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrackingLinkSentDto"
                }
              }
            }
          },
          "401": {
            "description": "UNAUTHENTICATED: no valid session",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "UNAUTHORIZED: signed in, but not allowed to do this",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "NOT_FOUND: the resource doesn't exist, or isn't visible to you",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "409": {
            "description": "CONFLICT: not possible in the resource's current state",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "429": {
            "description": "RATE_LIMITED: too many requests, retry after the Retry-After header's seconds",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          }
        },
        "summary": "Email the customer their tracking link again (ADR 0015)",
        "tags": [
          "Tickets"
        ]
      }
    },
    "/locations/{locationId}/tickets/{ticketId}/status/undo": {
      "post": {
        "operationId": "TicketController_undoStatus",
        "parameters": [
          {
            "name": "locationId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "ticketId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TicketDto"
                }
              }
            }
          },
          "401": {
            "description": "UNAUTHENTICATED: no valid session",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "UNAUTHORIZED: signed in, but not allowed to do this",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "NOT_FOUND: the resource doesn't exist, or isn't visible to you",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "409": {
            "description": "CONFLICT: not possible in the resource's current state",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "429": {
            "description": "RATE_LIMITED: too many requests, retry after the Retry-After header's seconds",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          }
        },
        "summary": "Undo your own latest status change, within 2 minutes (ADR 0016)",
        "tags": [
          "Tickets"
        ]
      }
    },
...
  },
  "info": {
    "title": "ReadyYet API",
    "description": "Better Auth's own routes (/api/auth/*) aren't included here, they're raw middleware, not Nest controllers. See ADR 0008/0011.",
    "version": "0.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "better-auth.session_token"
      }
    },
    "schemas": {
      "ValidationIssueDto": {
        "type": "object",
        "properties": {
          "constraints": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            }
          },
          "property": {
            "type": "string"
          }
        },
        "required": [
          "property"
        ]
      },
      "ApiErrorDto": {
        "type": "object",
        "properties": {
          "code": {
            "enum": [
              "VALIDATION_ERROR",
              "UNAUTHENTICATED",
              "UNAUTHORIZED",
              "REAUTHENTICATION_REQUIRED",
              "NOT_FOUND",
              "CONFLICT",
              "RATE_LIMITED",
              "INTERNAL_ERROR"
            ],
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "details": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ValidationIssueDto"
            }
          },
          "requestId": {
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ]
      },
      "ApiErrorResponseDto": {
        "type": "object",
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ApiErrorDto"
          }
        },
        "required": [
          "error"
        ]
      },
...
      "TranslationDto": {
        "type": "object",
        "properties": {
          "locale": {
            "enum": [
              "EN",
              "FR"
            ],
            "type": "string"
          },
          "label": {
            "type": "string"
          }
        },
        "required": [
          "locale",
          "label"
        ]
      },
      "StatusDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "code": {
            "type": "string"
          },
          "translations": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TranslationDto"
            }
          }
        },
        "required": [
          "id",
          "code",
          "translations"
        ]
      },
...
      "CreateCustomerDto": {
        "type": "object",
        "properties": {
          "fullName": {
            "type": "string"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "phone": {
            "type": "string"
          },
          "locale": {
            "type": "string",
            "enum": [
              "EN",
              "FR"
            ]
          }
        },
        "required": [
          "fullName"
        ]
      },
      "CreateTicketDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "customerId": {
            "type": "string"
          },
          "customer": {
            "$ref": "#/components/schemas/CreateCustomerDto"
          }
        },
        "required": [
          "title"
        ]
      },
      "CustomerDto": {
        "type": "object",
        "properties": {
          "locale": {
            "nullable": true,
            "enum": [
              "EN",
              "FR"
            ],
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "fullName": {
            "type": "string"
          },
          "email": {
            "type": "string",
            "nullable": true
          },
          "phone": {
            "type": "string",
            "nullable": true
          },
          "emailBouncedAt": {
            "format": "date-time",
            "type": "string",
            "nullable": true
          },
          "emailComplainedAt": {
            "format": "date-time",
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "locale",
          "id",
          "fullName",
          "email",
          "phone",
          "emailBouncedAt",
          "emailComplainedAt",
          "createdAt"
        ]
      },
      "TicketDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "trackingCode": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "notificationsStoppedAt": {
            "format": "date-time",
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "customer": {
            "$ref": "#/components/schemas/CustomerDto"
          },
          "currentStatus": {
            "$ref": "#/components/schemas/StatusDto"
          }
        },
        "required": [
          "id",
          "trackingCode",
          "title",
          "description",
          "notificationsStoppedAt",
          "createdAt",
          "customer",
          "currentStatus"
        ]
      },
      "TicketListCustomerDto": {
        "type": "object",
        "properties": {
          "fullName": {
            "type": "string"
          }
        },
        "required": [
          "fullName"
        ]
      },
      "TicketListItemDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "trackingCode": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "customer": {
            "$ref": "#/components/schemas/TicketListCustomerDto"
          },
          "currentStatus": {
            "$ref": "#/components/schemas/StatusDto"
          }
        },
        "required": [
          "id",
          "trackingCode",
          "title",
          "createdAt",
          "customer",
          "currentStatus"
        ]
      },
      "TicketPageDto": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TicketListItemDto"
            }
          },
          "nextCursor": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "items",
          "nextCursor"
        ]
      },
      "TicketStatusEventDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "changedBy": {
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/StatusDto"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "changedBy",
          "status",
          "createdAt"
        ]
      },
      "TicketDetailDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "trackingCode": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "notificationsStoppedAt": {
            "format": "date-time",
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "customer": {
            "$ref": "#/components/schemas/CustomerDto"
          },
          "currentStatus": {
            "$ref": "#/components/schemas/StatusDto"
          },
          "statusEvents": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TicketStatusEventDto"
            }
          }
        },
        "required": [
          "id",
          "trackingCode",
          "title",
          "description",
          "notificationsStoppedAt",
          "createdAt",
          "customer",
          "currentStatus",
          "statusEvents"
        ]
      },
      "UpdateTicketDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          }
        }
      },
      "UpdateTicketStatusDto": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "string"
          }
        },
        "required": [
          "statusCode"
        ]
      },
      "TrackingLinkSentDto": {
        "type": "object",
        "properties": {
          "sentTo": {
            "type": "string"
          }
        },
        "required": [
          "sentTo"
        ]
      },
...
    }
  }
}